+++ lib/cuoral_flutter.dart
@@ -61,4 +61,4 @@ export 'cuoral_launcher.dart' show CuoralLauncher;
 export 'src/cuoral_overlay.dart' show CuoralOverlay;
 
 /// Version information
-const String version = '0.1.7';
+const String version = '0.1.8';